# openapi/owned-sales-automation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Owned Sales Automation API",
    "version": "0.1.0",
    "description": "Repo-owned API contract derived from the captured SalesAutomation swagger and extended with workflow endpoints required for assistant-driven automation."
  },
  "servers": [
    {
      "url": "/",
      "description": "Local application server"
    }
  ],
...
  "paths": {
...
    "/api/service-requests": {
      "get": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "List service requests visible to the current user.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated statuses to filter by."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestPagedResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "Create a new client or admin service request.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/service-requests/{requestId}/opportunity": {
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "Create or link an opportunity for a service request.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestOpportunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/service-requests/{requestId}/proposal": {
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "Create or link a proposal for a service request.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestProposalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalWithLineItemsDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/service-requests/{requestId}/client-decision": {
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "Apply the client's approval or rejection of proposed assignments.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/service-requests/{requestId}/rep-decision": {
      "post": {
        "tags": [
          "ServiceRequests"
        ],
        "summary": "Apply the representative's acceptance or rejection of an assignment.",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestRepDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "OpportunityStage": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "OpportunityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "contactId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerName": {
            "type": "string",
            "nullable": true
          },
          "estimatedValue": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "probability": {
            "type": "integer",
            "format": "int32"
          },
          "stage": {
            "$ref": "#/components/schemas/OpportunityStage"
          },
          "stageName": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/OpportunitySource"
          },
          "expectedCloseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualCloseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "lossReason": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProposalStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProposalWithLineItemsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "proposalNumber": {
            "type": "string",
            "nullable": true
          },
          "opportunityId": {
            "type": "string",
            "format": "uuid"
          },
          "opportunityTitle": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "clientName": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "statusName": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "validUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "submittedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "approvedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdByName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lineItemsCount": {
            "type": "integer",
            "format": "int32"
          },
          "lineItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalLineItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OpportunitySource": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProposalLineItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "productServiceName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "discount": {
            "type": "number",
            "format": "double"
          },
          "taxRate": {
            "type": "number",
            "format": "double"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ServiceRequestStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "under_review",
          "proposal_prepared",
          "awaiting_client_assignment_approval",
          "client_rejected_assignment",
          "client_approved_assignment",
          "awaiting_rep_acceptance",
          "rep_rejected_assignment",
          "active",
          "closed",
          "cancelled"
        ]
      },
      "ServiceRequestPriority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ]
      },
...
      "ServiceRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "submittedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "requestType": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceRequestStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/ServiceRequestPriority"
          },
          "source": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "clientId",
          "submittedByUserId",
          "title",
          "description",
          "requestType",
          "status",
          "priority",
          "source",
          "createdAt",
          "updatedAt"
        ]
      },
      "ServiceRequestPagedResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceRequestDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateServiceRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "requestType": {
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/ServiceRequestPriority"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "clientId",
          "title",
          "description"
        ]
      },
...
      "ServiceRequestOpportunityDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "create",
              "link"
            ]
          },
          "opportunityId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "estimatedValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "expectedCloseDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "stage": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "mode"
        ]
      },
      "ServiceRequestProposalDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "create",
              "link"
            ]
          },
          "proposalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "validUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "mode"
        ]
      },
...
      "ServiceRequestDecisionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ]
          },
          "assignmentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "decision",
          "assignmentIds"
        ]
      },
      "ServiceRequestRepDecisionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "assignmentId": {
            "type": "string"
          },
          "decision": {
            "type": "string",
            "enum": [
              "accept",
              "reject"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "assignmentId",
          "decision"
        ]
      }
    }
  }
}